# apps/go/api-gateway/internal/apidocs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BPMP Platform Public API",
    "version": "1.0.0",
    "description": "Authoritative public HTTP contract for workflow commands, human work, audit queries, and runtime configuration lifecycle.",
    "license": {"name": "Apache-2.0", "identifier": "Apache-2.0"}
  },
  "servers": [{"url": "/"}],
  "security": [{"bearerAuth": []}],
...
  "paths": {
    "/v1/runtime/browser-configuration": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantID"},
        {"$ref": "#/components/parameters/CorrelationID"}
      ],
      "get": {
        "summary": "Resolve browser-safe tenant configuration",
        "operationId": "getBrowserConfiguration",
        "tags": ["Configuration"],
        "responses": {
          "200": {
            "description": "Browser-safe resolved configuration",
            "headers": {"ETag": {"schema": {"type": "string"}}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BrowserConfiguration"}}}
          },
          "304": {"description": "Configuration version has not changed"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "429": {"$ref": "#/components/responses/RateLimited"},
          "502": {"$ref": "#/components/responses/UpstreamUnavailable"}
        }
      }
    },
    "/v1/workflows/{workflowType}/instances": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantID"},
        {"$ref": "#/components/parameters/CorrelationID"},
        {"$ref": "#/components/parameters/CommandID"},
        {"$ref": "#/components/parameters/IdempotencyKey"},
        {"name": "workflowType", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/Identifier"}}
      ],
      "post": {
        "summary": "Start a workflow instance",
        "operationId": "startWorkflow",
        "tags": ["Workflow"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StartWorkflowRequest"}}}},
        "responses": {
          "202": {"description": "Command durably committed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CommandReceipt"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "429": {"$ref": "#/components/responses/RateLimited"},
          "502": {"$ref": "#/components/responses/UpstreamUnavailable"}
        }
      }
    },
    "/v1/work-items": {
      "parameters": [{"$ref": "#/components/parameters/TenantID"}, {"$ref": "#/components/parameters/CorrelationID"}],
      "get": {
        "summary": "List work items",
        "operationId": "listWorkItems",
        "tags": ["Human Work"],
        "parameters": [{"$ref": "#/components/parameters/PageSize"}, {"$ref": "#/components/parameters/PageToken"}],
        "responses": {
          "200": {"description": "Work-item page", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListWorkItemsResponse"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "429": {"$ref": "#/components/responses/RateLimited"},
          "502": {"$ref": "#/components/responses/UpstreamUnavailable"}
        }
      }
    },
    "/v1/work-items/{workItemID}": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantID"},
        {"$ref": "#/components/parameters/CorrelationID"},
        {"name": "workItemID", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/Identifier"}}
      ],
      "get": {
        "summary": "Get a work item",
        "operationId": "getWorkItem",
        "tags": ["Human Work"],
        "responses": {
          "200": {"description": "Work item", "content": {"application/json": {"schema": {"type": "object", "required": ["work_item"], "properties": {"work_item": {"$ref": "#/components/schemas/WorkItem"}}}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "502": {"$ref": "#/components/responses/UpstreamUnavailable"}
        }
      }
    },
    "/v1/work-items/{workItemID}/complete": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantID"}, {"$ref": "#/components/parameters/CorrelationID"},
        {"$ref": "#/components/parameters/CommandID"}, {"$ref": "#/components/parameters/IdempotencyKey"},
        {"name": "workItemID", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/Identifier"}}
      ],
      "post": {
        "summary": "Complete a work item",
        "operationId": "completeWorkItem",
        "tags": ["Human Work"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CompleteWorkItemRequest"}}}},
        "responses": {
          "202": {"description": "Completion accepted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkItemCommandResponse"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"}, "403": {"$ref": "#/components/responses/Forbidden"},
          "429": {"$ref": "#/components/responses/RateLimited"}, "502": {"$ref": "#/components/responses/UpstreamUnavailable"}
        }
      }
    },
    "/v1/work-items/{workItemID}/delegate": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantID"}, {"$ref": "#/components/parameters/CorrelationID"},
        {"$ref": "#/components/parameters/CommandID"}, {"$ref": "#/components/parameters/IdempotencyKey"},
        {"name": "workItemID", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/Identifier"}}
      ],
      "post": {
        "summary": "Delegate a work item",
        "operationId": "delegateWorkItem",
        "tags": ["Human Work"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DelegateWorkItemRequest"}}}},
        "responses": {
          "200": {"description": "Delegation committed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkItemCommandResponse"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"}, "403": {"$ref": "#/components/responses/Forbidden"},
          "429": {"$ref": "#/components/responses/RateLimited"}, "502": {"$ref": "#/components/responses/UpstreamUnavailable"}
        }
      }
    },
    "/v1/cases/{caseID}": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantID"}, {"$ref": "#/components/parameters/CorrelationID"},
        {"name": "caseID", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/Identifier"}}
      ],
      "get": {
        "summary": "Get a case",
        "operationId": "getCase",
        "tags": ["Human Work"],
        "responses": {
          "200": {"description": "Case view", "content": {"application/json": {"schema": {"type": "object", "required": ["case"], "properties": {"case": {"$ref": "#/components/schemas/Case"}}}}}},
          "400": {"$ref": "#/components/responses/BadRequest"}, "403": {"$ref": "#/components/responses/Forbidden"},
          "502": {"$ref": "#/components/responses/UpstreamUnavailable"}
        }
      }
    },
    "/v1/audit-records": {
      "parameters": [{"$ref": "#/components/parameters/TenantID"}, {"$ref": "#/components/parameters/CorrelationID"}],
      "get": {
        "summary": "List immutable audit records",
        "operationId": "listAuditRecords",
        "tags": ["Audit"],
        "parameters": [
          {"$ref": "#/components/parameters/PageSize"}, {"$ref": "#/components/parameters/PageToken"},
          {"name": "work_item_id", "in": "query", "schema": {"$ref": "#/components/schemas/Identifier"}},
          {"name": "case_id", "in": "query", "schema": {"$ref": "#/components/schemas/Identifier"}}
        ],
        "responses": {
          "200": {"description": "Immutable audit page", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListAuditRecordsResponse"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"}, "403": {"$ref": "#/components/responses/Forbidden"},
          "502": {"$ref": "#/components/responses/UpstreamUnavailable"}
        }
      }
    },
    "/v1/configuration/profiles": {
      "parameters": [{"$ref": "#/components/parameters/TenantID"}, {"$ref": "#/components/parameters/CorrelationID"}],
      "get": {
        "summary": "List configuration profiles",
        "operationId": "listConfigurationProfiles",
        "tags": ["Configuration"],
        "parameters": [{"$ref": "#/components/parameters/PageSize"}, {"$ref": "#/components/parameters/PageToken"}],
        "responses": {"200": {"description": "Configuration profile page", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigurationProfilePage"}}}}, "403": {"$ref": "#/components/responses/Forbidden"}, "502": {"$ref": "#/components/responses/UpstreamUnavailable"}}
      },
      "post": {
        "summary": "Create a configuration profile",
        "operationId": "createConfigurationProfile",
        "tags": ["Configuration"],
        "parameters": [{"$ref": "#/components/parameters/CommandID"}, {"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateConfigurationProfileRequest"}}}},
        "responses": {"201": {"description": "Profile created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigurationProfile"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "403": {"$ref": "#/components/responses/Forbidden"}, "502": {"$ref": "#/components/responses/UpstreamUnavailable"}}
      }
    },
    "/v1/configuration/profiles/{profileID}": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantID"}, {"$ref": "#/components/parameters/CorrelationID"},
        {"name": "profileID", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/Identifier"}}
      ],
      "get": {
        "summary": "Get a configuration profile and its versions",
        "operationId": "getConfigurationProfile",
        "tags": ["Configuration"],
        "responses": {"200": {"description": "Profile and version history", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigurationProfileDetail"}}}}, "403": {"$ref": "#/components/responses/Forbidden"}, "404": {"$ref": "#/components/responses/NotFound"}, "502": {"$ref": "#/components/responses/UpstreamUnavailable"}}
      }
    },
    "/v1/configuration/profiles/{profileID}/versions": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantID"}, {"$ref": "#/components/parameters/CorrelationID"},
        {"$ref": "#/components/parameters/CommandID"}, {"$ref": "#/components/parameters/IdempotencyKey"},
        {"name": "profileID", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/Identifier"}}
      ],
      "post": {
        "summary": "Add a draft configuration version",
        "operationId": "addConfigurationDraft",
        "tags": ["Configuration"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AddConfigurationDraftRequest"}}}},
        "responses": {"201": {"description": "Draft version created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigurationProfile"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "403": {"$ref": "#/components/responses/Forbidden"}, "409": {"$ref": "#/components/responses/Conflict"}, "502": {"$ref": "#/components/responses/UpstreamUnavailable"}}
      }
    },
    "/v1/configuration/profiles/{profileID}/versions/{versionID}/publish": {
      "parameters": [{"$ref": "#/components/parameters/TenantID"}, {"$ref": "#/components/parameters/CorrelationID"}, {"$ref": "#/components/parameters/CommandID"}, {"$ref": "#/components/parameters/IdempotencyKey"}, {"$ref": "#/components/parameters/ProfileID"}, {"$ref": "#/components/parameters/VersionID"}],
      "post": {
        "summary": "Publish a configuration version",
        "operationId": "publishConfigurationVersion",
        "tags": ["Configuration"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigurationTransitionRequest"}}}},
        "responses": {"200": {"description": "Version published", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigurationProfile"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "403": {"$ref": "#/components/responses/Forbidden"}, "409": {"$ref": "#/components/responses/Conflict"}, "502": {"$ref": "#/components/responses/UpstreamUnavailable"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}},
    "parameters": {
      "TenantID": {"name": "X-BPMP-Tenant-ID", "in": "header", "required": true, "schema": {"$ref": "#/components/schemas/Identifier"}},
      "CorrelationID": {"name": "X-Correlation-ID", "in": "header", "required": true, "schema": {"$ref": "#/components/schemas/Identifier"}},
      "CommandID": {"name": "X-Command-ID", "in": "header", "required": true, "schema": {"$ref": "#/components/schemas/Identifier"}},
      "IdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 256}},
...
    },
    "responses": {
...
      "Forbidden": {"description": "Authentication or tenant authorization failed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
      "RateLimited": {"description": "Rate limit exceeded", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "UpstreamUnavailable": {"description": "Authoritative service unavailable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
    },
    "schemas": {
      "Identifier": {"type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[A-Za-z0-9_.:-]+$"},
      "BrowserConfiguration": {
        "type": "object",
        "additionalProperties": false,
        "required": ["config_version", "policy_version", "batch_chunk_size", "batch_concurrency"],
        "properties": {
          "config_version": {"type": "string", "minLength": 1},
          "policy_version": {"type": "string", "minLength": 1},
          "batch_chunk_size": {"type": "integer", "minimum": 1},
          "batch_concurrency": {"type": "integer", "minimum": 1}
        }
      },
...
    }
  }
}